Split handlers into multiple folders; update gencode to deal with this
