+++ src/main/handlers/media/thumbnailExtract.ts
