+++ src/main/handlers/misc/getFileNameWithExtension.ts
