+++ src/main/handlers/window/showConfirmation.ts
@@ -1,5 +1,5 @@
 import { dialog } from 'electron';
-import { IpcContext } from '../types';
+import { IpcContext } from '../../types';
 
 type ShowConfirmation = (
   ipcContext: IpcContext,
